add a lot of blog features 2 and fix bugs

+++ frontend/src/containers/BookmarkContainer/BookmarkContainer.tsx
@@ -1,12 +1,14 @@
 import NcBookmark, { NcBookmarkProps } from "components/NcBookmark/NcBookmark";
 import React from "react";
 import { useAppSelector, useAppDispatch } from "app/hooks";
+import { useHistory } from "react-router-dom";
 import {
   addNewSavedByPostId,
   removeSavedByPostId,
   selectRecentSaveds,
   selectRecentRemoveds,
 } from "app/bookmarks/bookmarksSlice";
+import { selectAuthState } from "app/auth/auth";
 
 export type BookmarkContainerProps = Omit<NcBookmarkProps, "isBookmarked"> & {
   initBookmarked: boolean;
@@ -16,7 +18,9 @@ const BookmarkContainer: React.FC<BookmarkContainerProps> = (props) => {
   const { postId, initBookmarked } = props;
   const recentSaveds = useAppSelector(selectRecentSaveds);
   const recentRemoveds = useAppSelector(selectRecentRemoveds);
+  const auth = useAppSelector(selectAuthState);
   const dispatch = useAppDispatch();
+  const history = useHistory();
   const isBookmarked = () => {
     if (recentSaveds.includes(postId)) {
       return true;
@@ -28,6 +32,13 @@ const BookmarkContainer: React.FC<BookmarkContainerProps> = (props) => {
   };
 
   const handleClickBookmark = () => {
+    if (!auth.token) {
+      history.push({
+        pathname: "/login",
+        state: { from: history.location.pathname } // Save the current path as the "from" state
+      });
+      return;
+    }
     if (isBookmarked()) {
       dispatch(removeSavedByPostId(postId));
     } else {
